+++ changelog.md
@@ -1,4 +1,26 @@
-### 2020 April 9
+
+## 2020-04-10
+
+** BREAKING CHANGES **
+
+* add argparse to enable proper command line switches
+
+```
+usage: check_all_by_postal_code.py [-h] [-p P] [-lat LAT] [-long LONG] [-d D] [-r]
+
+optional arguments:
+  -h, --help         show this help message and exit
+  -p P, -postal P    ex: m5w1e6
+  -lat LAT           ex: 43.703344
+  -long LONG         ex: -79.524619
+  -d D, -distance D  Search distance in KM
+  -r, -report        report lat + long and stores found within search distance. will not check available pickup times
+```
+* remove MYLAT and MYLONG environment variables
+
+* prefer `-lat` and `-long` over `-p` postal code
+
+## 2020-04-09
 
 * check for MYLAT and MYLONG environment variables to bypass geo lookup request. If the report option is used, it will give you the variables to use along with the locations.
 ```bash
@@ -13,7 +35,7 @@ store: nofrills, location: 3480, address: 245 Dixon Rd Etobicoke, Ontario M9P 2M
 
 * additional filtering of locations.json
 
-### 2020 April 8
+## 2020-04-08
 
 * fallback to NRC geo lookup in the event that geocode.ca throttles requests
 
@@ -23,7 +45,7 @@ store: nofrills, location: 3480, address: 245 Dixon Rd Etobicoke, Ontario M9P 2M
 
 * add secondary postal code lookup since the NRC geo lookup may return nothing
 
-### 2020 April 6
+## 2020-04-06
 
 * add Dockerfile and link to dockerhub
 
@@ -35,11 +57,11 @@ store: nofrills, location: 3480, address: 245 Dixon Rd Etobicoke, Ontario M9P 2M
 
 * change base_url to a dictionary lookup
 
-### 2020 April 5
+## 2020-04-05
 
 * merge [Lookup by PostCode / Specify Max Dist in KM feature](https://github.com/shmick/pcexpress-pickup/pull/1) ( Big thanks to [adam1](https://github.com/1adam) )
 
-### 2020 April 1
+## 2020-04-01
 
 * add LICENSE